Clarify that this research is for The Grid project, not Matrix
We originally choice to remain neutral on the origin of the document
and to be used for The Grid Protocol project, as Matrix.org leadership
usually see us as a hostile fork. We therefore wanted to avoid giving
the impression we were promoting it at the expense of Matrix's shortcoming.

Matrix.org leadership gave feedback that it felt like a lack of disclosure
and some community members highlighted that the document could be read
in a more constructive light if that was mentionned.